pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "percos"
version = "0.1.0"
description = "Ontology-Governed Knowledge Base Framework – adaptive, self-developing, domain-agnostic"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.19",
    "chromadb>=0.4",
    "networkx>=3.2",
    "openai>=1.10",
    "tiktoken>=0.5",
    "httpx>=0.26",
    "python-dateutil>=2.8",
    "structlog>=24.1",
    "tenacity>=8.2",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
]
google = [
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
]
microsoft = [
    "msal>=1.24",
    "requests>=2.31",
]
all-integrations = [
    "percos[google,microsoft]",
]

[project.scripts]
percos = "percos.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
